add chart buttons
