+++ app/src/main/java/com/example/cashmate/ui/category/BarChartActivity.kt
@@ -1,0 +1,12 @@
+package com.example.cashmate.ui.category
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.cashmate.R
+
+class BarChartActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_bar_chart)
+    }
+}
