+++ app/src/main/java/com/example/cashmate/ui/category/PieChartActivity.kt
@@ -1,0 +1,12 @@
+package com.example.cashmate.ui.category
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.cashmate.R
+
+class PieChartActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_pie_chart)
+    }
+}
