+++ app/src/main/java/com/example/cashmate/ui/category/SlideshowFragment.kt
@@ -1,0 +1,38 @@
+package com.example.cashmate.ui.category
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.example.cashmate.databinding.FragmentSlideshowBinding
+
+class SlideshowFragment : Fragment() {
+
+    private var _binding: FragmentSlideshowBinding? = null
+
+    // This property is only valid between onCreateView and
+    // onDestroyView.
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val slideshowViewModel =
+            ViewModelProvider(this).get(SlideshowViewModel::class.java)
+
+        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
+        val root: View = binding.root
+
+        return root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
